+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/complaint/ComplaintManager.java
@@ -188,7 +188,7 @@ public class ComplaintManager extends SecuredManager {
     public ChplOneTimeTrigger triggerComplaintsReport() throws SchedulerException, ValidationException {
         UserDTO jobUser = null;
         try {
-            jobUser = userManager.getById((Long) AuthUtil.getCurrentUser().getId());
+            jobUser = userManager.getById(AuthUtil.getCurrentUser().getId());
         } catch (UserRetrievalException ex) {
             LOGGER.error("Could not find user to execute job.");
         }
